refactoring app into components and added data folder
